+++ src_tests/test_Tridiagonal.cc
@@ -127,8 +127,8 @@ void print_test_table( const std::vector<TestResult> & results, const std::strin
     "└──────────────┴──────────────┴──────────────────────┴────────────┴───────────┘\n" );
 
   // Summary for this table
-  int passed = std::count_if( results.begin(), results.end(), []( const TestResult & r ) { return r.passed; } );
-  int total  = results.size();
+  int passed = static_cast<int>( std::count_if( results.begin(), results.end(), []( const TestResult & r ) { return r.passed; } ) );
+  int total  = static_cast<int>( results.size() );
 
   fmt::print(
     fg( passed == total ? fmt::color::green : fmt::color::red ) | fmt::emphasis::bold,
